+++ CataloguesAnywhere/Framework/ShopIds.cs
@@ -1,0 +1,26 @@
+namespace CataloguesAnywhere.Framework
+{
+    internal static class ShopIds
+    {
+        /// <summary>The shop ID for the generic furniture catalogue.</summary>
+        public const string FurnitureCatalogueId = "Furniture Catalogue";
+
+        /// <summary>The shop ID for the Joja furniture catalogue.</summary>
+        public const string JojaFurnitureCatalogueId = "JojaFurnitureCatalogue";
+
+        /// <summary>The shop ID for the Junimo furniture catalogue.</summary>
+        public const string JunimoFurnitureCatalogueId = "JunimoFurnitureCatalogue";
+
+        /// <summary>The shop ID for the Junimo furniture catalogue.</summary>
+        public const string RetroFurnitureCatalogueId = "RetroFurnitureCatalogue";
+
+        /// <summary>The shop ID for the Trash furniture catalogue.</summary>
+        public const string TrashFurnitureCatalogueId = "TrashFurnitureCatalogue";
+
+        /// <summary>The shop ID for the Wizard furniture catalogue.</summary>
+        public const string WizardFurnitureCatalogueId = "WizardFurnitureCatalogue";
+
+        /// <summary>The shop ID for the wallpaper/flooring catalogue.</summary>
+        public const string WallpaperCatalogueId = "Catalogue";
+    }
+}