+++ CataloguesAnywhere/ModEntry.cs
@@ -2,7 +2,9 @@ using CataloguesAnywhere.Framework;
 using ChroniclerCherry.Common.Integrations.GenericModConfigMenu;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
+using StardewModdingAPI.Utilities;
 using StardewValley;
+using StardewValley.Menus;
 
 namespace CataloguesAnywhere;
 
@@ -37,7 +39,7 @@ internal class ModEntry : Mod
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
         this.AddGenericModConfigMenu(
-            new GenericModConfigMenuIntegrationForCataloguesAnywhere(),
+            new GenericModConfigMenuIntegrationForCataloguesAnywhere(() => this.Config),
             get: () => this.Config,
             set: config => this.Config = config
         );
@@ -46,12 +48,38 @@ internal class ModEntry : Mod
     /// <inheritdoc cref="IInputEvents.ButtonsChanged" />
     private void OnButtonsChanged(object? sender, ButtonsChangedEventArgs e)
     {
-        if (!Context.CanPlayerMove || !this.Config.Enabled)
+        if (!Context.IsWorldReady || !this.Config.Enabled)
             return;
 
-        if (this.Config.FurnitureKey.JustPressed())
-            Utility.TryOpenShopMenu("Furniture Catalogue", null as string);
-        else if (this.Config.WallpaperKey.JustPressed())
-            Utility.TryOpenShopMenu("Catalogue", null as string);
+        foreach ((string shopId, KeybindList keybind) in this.Config.Catalogues)
+        {
+            if (keybind.JustPressed())
+            {
+                if (this.CanOpenShopMenu(shopId))
+                {
+                    this.Monitor.Log($"Opening shop ID '{shopId}' per keybind '{keybind}'.");
+                    Utility.TryOpenShopMenu(shopId, null as string);
+                }
+
+                break;
+            }
+        }
+    }
+
+    /// <summary>Get whether a given shop menu can be opened now.</summary>
+    /// <param name="shopId">The shop ID being opened.</param>
+    private bool CanOpenShopMenu(string shopId)
+    {
+        return
+            // no menu open
+            Context.CanPlayerMove
+
+            // allow switching from one catalogue to another
+            || (
+                Game1.activeClickableMenu is ShopMenu { ShopId: not null } shopMenu
+                && shopMenu.ShopId != shopId
+                && this.Config.Catalogues.ContainsKey(shopMenu.ShopId)
+                && shopMenu.readyToClose()
+            );
     }
 }