+++ CataloguesAnywhere/docs/release-notes.md
@@ -1,4 +1,11 @@
 ﻿# Release notes
+## 1.3.0
+Released 11 January 2026 for SMAPI 4.1.10 or later. Updated by Pathoschild.
+
+- Added themed furniture catalogues in Stardew Valley 1.6+.
+- Added support for opening custom catalogue/shop IDs.
+- You can now switch catalogues with the keybinds while a catalogue is open.
+
 ## 1.2.0
 Released 07 September 2025 for SMAPI 4.1.10 or later. Updated by Pathoschild.